// src/app/core/models/register.model.ts
export interface RegisterModel {
  primerNombre: string | null;
  segundoNombre: string | null;
  primerApellido: string | null;
  segundoApellido: string | null;
  apellidoCasada: string | null;
  dpi: string | null;
  nit: string | null;
  pasaporte: string | null;
  telefono: string | null;
  correo: string | null;
  direccion: string | null;
  estado: string | null;
  categoria: string | null;
  idRol: number;
  contrasena: string | null;
}